[refactor] exporting Datastream time ranges checks into dedicated class
